git-svn-id: http://svn.purei.org/purei/trunk@179 b36208d7-6611-0410-8bec-b1987f11c4a2

+++ Source/X86Assembler.cpp
@@ -94,7 +94,7 @@ void CX86Assembler::ResolveLabelReferences()
             {
                 throw runtime_error("Label reference too small.");
             }
-            m_WriteAtFunction(referencePos, static_cast<uint8>(offset));
+            m_WriteAtFunction(static_cast<unsigned int>(referencePos), static_cast<uint8>(offset));
         }
     }
     m_labelReferences.clear();
